sync from original - 31 jan 2018

+++ src/game/Movement/util.cpp
@@ -21,7 +21,7 @@
 
 namespace Movement
 {
-    double gravity = 19.29110527038574;
+    float gravity = 19.29110527038574f;
 
     /// Velocity bounds that makes fall speed limited
     float terminalVelocity = 60.148003f;
